Decimal changed to double for higher precision, LoyaltyPointsRate added to dto

+++ SharedContracts/SharedContracts/DTOs/Location/Requests/RestaurantLocationData.cs
@@ -9,8 +9,8 @@ namespace PureDelivery.Shared.Contracts.DTOs.Location.Requests
     public class RestaurantLocationData
     {
         public Guid RestaurantId { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public List<DeliveryZoneData> DeliveryZones { get; set; } = new();
     }
 
@@ -31,7 +31,7 @@ namespace PureDelivery.Shared.Contracts.DTOs.Location.Requests
     public class ZonePointData
     {
         public int Order { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
